requirements.txt
flask
flask_session
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-dotenv